update with resourcedetaildrawer
