+++ src/components/content/care-team/patient-careteam.stories.tsx
@@ -1,6 +1,8 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { CareTeamDetailsDrawerProps } from "./careteam-details-drawer";
-import { PatientCareTeam } from "./patient-careteam";
+import {
+  CareTeamDetailsDrawerProps,
+  PatientCareTeam,
+} from "./patient-careteam";
 import { setupCareTeamMocks } from "./story-helpers/mocks/requests";
 import { CTWProvider } from "@/components/core/providers/ctw-provider";
 import { PatientProvider } from "@/components/core/providers/patient-provider";
